Add HTTPS test for Open Rosa POST
