+++ collect_app/src/test/java/org/odk/collect/android/http/OkHttpConnectionPostRequest.java
@@ -3,11 +3,21 @@ package org.odk.collect.android.http;
 import org.junit.runner.RunWith;
 import org.robolectric.RobolectricTestRunner;
 
+import okhttp3.OkHttpClient;
+import okhttp3.tls.internal.TlsUtil;
+
 @RunWith(RobolectricTestRunner.class)
 public class OkHttpConnectionPostRequest extends OpenRosaPostRequestTest {
 
     @Override
     protected OpenRosaHttpInterface buildSubject() {
-        return new OkHttpConnection(null);
+        return new OkHttpConnection(new OkHttpClient.Builder()
+                .sslSocketFactory(TlsUtil.localhost().sslSocketFactory(), TlsUtil.localhost().trustManager())
+        );
+    }
+
+    @Override
+    protected Boolean useRealHttps() {
+        return true;
     }
 }
