+++ collect_app/src/test/java/org/odk/collect/android/http/OpenRosaGetRequestTest.java
@@ -17,6 +17,7 @@ import java.util.zip.GZIPOutputStream;
 import okhttp3.mockwebserver.MockResponse;
 import okhttp3.mockwebserver.MockWebServer;
 import okhttp3.mockwebserver.RecordedRequest;
+import okhttp3.tls.internal.TlsUtil;
 import okio.Buffer;
 
 import static org.hamcrest.Matchers.equalTo;
@@ -27,8 +28,10 @@ import static org.junit.Assert.assertThat;
 public abstract class OpenRosaGetRequestTest {
 
     protected abstract OpenRosaHttpInterface buildSubject();
+    protected abstract Boolean useRealHttps();
 
     private final MockWebServer mockWebServer = new MockWebServer();
+    private MockWebServer httpsMockWebServer;
     private OpenRosaHttpInterface subject;
 
     @Before
@@ -40,6 +43,11 @@ public abstract class OpenRosaGetRequestTest {
     @After
     public void teardown() throws Exception {
         mockWebServer.shutdown();
+
+        if (httpsMockWebServer != null) {
+            httpsMockWebServer.shutdown();
+            httpsMockWebServer = null;
+        }
     }
 
     @Test
@@ -90,7 +98,7 @@ public abstract class OpenRosaGetRequestTest {
     }
 
     @Test
-    public void withCredentials_whenHttp_doesNotRetryWithCredentials() throws Exception  {
+    public void withCredentials_whenHttp_doesNotRetryWithCredentials() throws Exception {
         mockWebServer.enqueue(new MockResponse()
                 .setResponseCode(401)
                 .addHeader("WWW-Authenticate: Basic realm=\"protected area\"")
@@ -103,9 +111,27 @@ public abstract class OpenRosaGetRequestTest {
     }
 
     @Test
+    public void withCredentials_whenHttps_retriesWithCredentials() throws Exception {
+        startHttpsMockWebServer();
+
+        httpsMockWebServer.enqueue(new MockResponse()
+                .setResponseCode(401)
+                .addHeader("WWW-Authenticate: Basic realm=\"protected area\"")
+                .setBody("Please authenticate."));
+        httpsMockWebServer.enqueue(new MockResponse());
+
+        subject.executeGetRequest(httpsMockWebServer.url("").uri(), null, new HttpCredentials("user", "pass"));
+
+        assertThat(httpsMockWebServer.getRequestCount(), equalTo(2));
+        httpsMockWebServer.takeRequest();
+        RecordedRequest request = httpsMockWebServer.takeRequest();
+        assertThat(request.getHeader("Authorization"), equalTo("Basic dXNlcjpwYXNz"));
+    }
+
+    @Test
     public void whenLastRequestSetCookies_nextRequestDoesNotSendThem() throws Exception {
         mockWebServer.enqueue(new MockResponse()
-            .addHeader("Set-Cookie", "blah=blah"));
+                .addHeader("Set-Cookie", "blah=blah"));
         mockWebServer.enqueue(new MockResponse());
 
         subject.executeGetRequest(mockWebServer.url("").uri(), null, null);
@@ -201,6 +227,16 @@ public abstract class OpenRosaGetRequestTest {
         HttpGetResult result2 = subject.executeGetRequest(mockWebServer.url("").uri(), null, null);
         assertThat(result2.getInputStream(), nullValue());
         assertThat(result2.getStatusCode(), equalTo(304));
+    }
+
+    private void startHttpsMockWebServer() throws IOException {
+        httpsMockWebServer = new MockWebServer();
+
+        if (useRealHttps()) {
+            httpsMockWebServer.useHttps(TlsUtil.localhost().sslSocketFactory(), false);
+        }
+
+        httpsMockWebServer.start(8443);
     }
 
     private static byte[] gzip(String data) throws IOException {
